Beginning of Kubernetes integration
This is probably going to need to evolve into a full controller with CRDs for
each SRV to resolve. Realistically, a simple control group doesn't work and the
domains being passed into the config for the controller isn't reasonable or
scalable.

This will be a bit of an overhaul and a deviation from the simple "create a
service and endpoint" approach up to this point.